Implemented Updates Per Second inside the main Application. Also seperated Updates and Renders within the Layer class. Finally, seperated Updates and Swap Buffers within the Window class.

+++ src/Ares/Engine/Core/Timestep.h
@@ -7,17 +7,17 @@ namespace Ares {
 	class Timestep
 	{
 	public:
-		Timestep(float time = 0.0f)
+		Timestep(double time = 0.0f)
 			: m_Time(time)
 		{
 		}
 
-		operator float() const { return m_Time; }
+		operator double() const { return m_Time; }
 
-		float GetSeconds() const { return m_Time; }
-		float GetMilliseconds() const { return m_Time * 1000.0f; }
+		double GetSeconds() const { return m_Time; }
+		double GetMilliseconds() const { return m_Time * 1000.0f; }
 
 	private:
-		float m_Time;
+		double m_Time;
 	};
 }
